Refactorig gee_upload.py, pullout catalog code.
